shared/pyproject.toml
[project]
name = "shared-kernel"
version = "2.0.0"
description = "DDD Shared Kernel for garbage classification system"
requires-python = ">=3.10"

dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jinja2>=3.0",
]
